requirements.txt
fastapi>=0.112.0
uvicorn[standard]>=0.30.0
SQLAlchemy>=2.0.25
alembic>=1.13.0
psycopg[binary]>=3.1.19
pydantic>=2.7.0
pydantic-core>=2.18.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
httpx>=0.27.0
firebase-admin>=6.5.0
resend>=0.7.0
svix>=1.9.0
python-multipart>=0.0.9
pytest>=8.2.0
python-dotenv>=1.0.0